refactor(channelPage): 글 상세 페이지 커스텀 훅/컴포넌트 분리 - 댓글 드롭다운 컴포넌트 분리 - 게시글 정보 가져오는 로직 분리 - 댓글 목록 관련 로직 분리
